Created basic file format for compressed data

+++ FractalImageCompression/Block.cs
@@ -9,7 +9,7 @@ namespace FractalImageCompression
         public Atom AtomicData { get; }
         public Image Image { get; }
 
-        public Block(int x, int y, Transform transform, Image image)
+        public Block(byte x, byte y, Transform transform, Image image)
         {
             AtomicData = new Atom(x, y, 1.0f, 0.0f, transform);
             Image = image;
